Refactor: change TypicalIndexes to TypicalPersonIndexes
