+++ src/test/java/swift/logic/commands/EditCommandTest.java
@@ -10,8 +10,8 @@ import static swift.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 import static swift.logic.commands.CommandTestUtil.assertCommandFailure;
 import static swift.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static swift.logic.commands.CommandTestUtil.showPersonAtIndex;
-import static swift.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
-import static swift.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
+import static swift.testutil.TypicalPersonIndexes.INDEX_FIRST_PERSON;
+import static swift.testutil.TypicalPersonIndexes.INDEX_SECOND_PERSON;
 import static swift.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
