+++ src/test/java/swift/logic/commands/ListCommandTest.java
@@ -2,7 +2,7 @@ package swift.logic.commands;
 
 import static swift.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static swift.logic.commands.CommandTestUtil.showPersonAtIndex;
-import static swift.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
+import static swift.testutil.TypicalPersonIndexes.INDEX_FIRST_PERSON;
 import static swift.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.BeforeEach;
