+++ src/test/java/swift/testutil/TypicalPersonIndexes.java
@@ -5,7 +5,7 @@ import swift.commons.core.index.Index;
 /**
  * A utility class containing a list of {@code Index} objects to be used in tests.
  */
-public class TypicalIndexes {
+public class TypicalPersonIndexes {
     public static final Index INDEX_FIRST_PERSON = Index.fromOneBased(1);
     public static final Index INDEX_SECOND_PERSON = Index.fromOneBased(2);
     public static final Index INDEX_THIRD_PERSON = Index.fromOneBased(3);
